+++ app/abs/[id]/page.js
@@ -6,58 +6,13 @@ import Metadata from '@/app/components/Metadata';
 import LoadingState from '@/app/components/LoadingState';
 import { Toaster, toast } from 'react-hot-toast';
 import Layout from '@/app/components/Layout';
+import { usePaperData } from '@/app/hooks/usePaperData';
+
 export default function PaperPage({ params }) {
   const unwrappedParams = use(params);
   const { id } = unwrappedParams;
-  const [paper, setPaper] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-  const [plainTextContent, setPlainTextContent] = useState('');
+  const { paper, loading, error, plainTextContent } = usePaperData(id);
 
-  useEffect(() => {
-    async function fetchPaper() {
-      setLoading(true);
-      setError(null);
-
-      try {
-        const response = await fetch(`/raw/abs/${id}`);
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch paper: ${response.status} ${response.statusText}`);
-        }
-
-        const content = await response.text();
-        setPlainTextContent(content);
-
-        // Parse the plain text content to extract structured data
-        const sections = content.split('\n\n');
-        const paperData = {
-          title: sections[0].replace('# ', ''),
-          authors: sections[1].replace('## Authors\n', '').split(', '),
-          categories: sections[2].replace('## Categories\n', '').split(', '),
-          abstract: sections[sections.length - 1].replace('## Abstract\n', ''),
-          id: id
-        };
-
-        // Extract DOI if present
-        const doiMatch = content.match(/DOI: (.*)/);
-        if (doiMatch) {
-          paperData.doi = doiMatch[1];
-        }
-
-        setPaper(paperData);
-      } catch (err) {
-        console.error('Error loading paper:', err);
-        setError(err.message || 'Failed to load paper information');
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    if (id) {
-      fetchPaper();
-    }
-  }, [id]);
 
   if (loading) {
     return <LoadingState />;
